% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propvacc_funcs.R
\name{calc_prop_vacc}
\alias{calc_prop_vacc}
\title{Calculate the total proportion vaccinated by routine vaccination}
\usage{
calc_prop_vacc(V, effectiveness, independent = FALSE)
}
\arguments{
\item{V}{a vector giving the proportion vaccinated for up to three routine immunization activities}

\item{effectiveness}{scalar or vector giving the vaccine effectiveness for each number of doses}

\item{independent}{logical indicating if receipt of routine vaccine dose is depends on the number
of prior doses received (default = FALSE)}
}
\value{
A scalar giving the total proportion of the population immune due to vaccination
}
\description{
This function calculates the proportion immune due to vaccination given the proportion vaccinated at
each routine activity and the effectiveness of each dose. The function assumes that vaccination events
are dependent by default, where individuals that have recieved the first
dose are the most likely to recieve the second dose and those that have received both the first
and second doses are the most likely to receive the third. The function uses either the two-dose
or three-dose method based on the length of \code{V}. Under the assumption of dependence, the total
proportion vaccinated is:
\deqn{
p_{\text{vacc}} = \sum_j \text{efficacy}_j \times \Pr(\text{dose}_j)
}{
 p_vacc = \sum_j effectiveness_j * Pr(dose_j)
}
Under the assumption of independence, the total proportion vaccinated is:
 \deqn{
p_{\text{vacc}} = 1 - ( \prod_j effectiveness_j * \Pr(not vaccinated by V_j) )
}{
 p_vacc = 1 - [ \prod_j effectiveness_j * Pr(not vaccinated by V_j) ]
}
}
\examples{
# Two dose measles vaccine: routine without SIA
calc_prop_vacc(V=c(0.9, 0.8),
               effectiveness=c(0.84, 0.941),
               independent=FALSE)

calc_prop_vacc(V=c(0.9, 0.8),
               effectiveness=c(0.84, 0.941),
               independent=TRUE)

# Three dose vaccine: routine without SIA
calc_prop_vacc(V=c(0.9, 0.8, 0.7),
               effectiveness=c(0.85, 0.9, 0.95))

# Estimate posterior distribution of proportion vaccinated given uncertainty around MCV1 and MCV2
#sims <- foreach(i=1:5000, .combine='c') \%do\% {

#  p <- calc_prop_vacc(V=c(rbeta(1,40,1), rbeta(1,4,2)),
#                      effectiveness=c(0.85, 0.94),
#                      independent = F)
#}

#q <- quantile(sims, c(0.025, 0.5, 0.975))

#par(mfrow=c(1,1))
#hist(sims, breaks=100, col='cyan', xlab='Proportion vaccinated')
#abline(v=q[2], lwd=3)
#abline(v=q[c(1,3)], lty=2, lwd=2)

}
\seealso{
Other prop_vacc: 
\code{\link{calc_doses_SIA}()},
\code{\link{calc_doses}()},
\code{\link{calc_prop_vacc_SIA}()}
}
\author{
John Giles
}
\concept{prop_vacc}
