% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{calc_prop_vacc_SIA}
\alias{calc_prop_vacc_SIA}
\title{Calculate the total proportion vaccinated by routine vaccination and SIA campaign}
\usage{
calc_prop_vacc_SIA(V, S, effectiveness, independent = FALSE)
}
\arguments{
\item{V}{a vector giving the proportion vaccinated for up to three routine immunization activities}

\item{S}{a scalar giving the proportion vaccinated with SIA campaign}

\item{effectiveness}{scalar or vector giving the vaccine effectiveness for each number of doses}

\item{independent}{logical indicating if receipt of routine vaccine dose is depends on the number
of prior doses received and receipt of SIA dose depends on at least one prior dose (default = FALSE)}
}
\value{
A scalar giving the total proportion of the population immune due to vaccination
}
\description{
This function calculates the proportion immune due to vaccination given the proportion vaccinated at
each routine activity plus one SIA campaign and the effectiveness of each dose. The function assumes
that vaccination events are dependent by default, where individuals that have recieved the first
dose are the most likely to recieve the second dose and those that have received both the first
and second doses are the most likely to receive the third. receipt of dose in SIA campaign is dependent on
having any number of prior doses. The function uses either the two-dose plus SIA
or three-dose plus SIA method based on the length of \code{V}.
}
\details{
Length of effectiveness must be equal to the number of vaccination activities
}
\examples{
# Two dose measles vaccine: routine without SIA
calc_prop_vacc(V=c(0.9, 0.8),
               effectiveness=c(0.84, 0.941),
               independent=FALSE)

calc_prop_vacc(V=c(0.9, 0.8),
               effectiveness=c(0.84, 0.941),
               independent=TRUE)

# Three dose vaccine: routine without SIA
calc_prop_vacc(V=c(0.9, 0.8, 0.7),
               effectiveness=c(0.85, 0.9, 0.95))

# Estimate posterior distribution of proportion vaccinated given uncertainty around MCV1 and MCV2
n <- 1000
sims <- rep(NA, n)
for (i in 1:n) {

  sims[i] <- calc_prop_vacc(V=c(rbeta(1,40,1), rbeta(1,4,2)),
                      effectiveness=c(0.85, 0.94),
                      independent=FALSE)
}

q <- quantile(sims, c(0.025, 0.5, 0.975))

par(mfrow=c(1,1))
hist(sims, breaks=100, col='cyan', xlab='Proportion vaccinated')
abline(v=q[2], lwd=3)
abline(v=q[c(1,3)], lty=2, lwd=2)

}
\seealso{
Other prop_vacc: 
\code{\link{calc_doses_SIA}()},
\code{\link{calc_doses}()},
\code{\link{calc_prop_vacc}()}
}
\author{
John Giles
}
\concept{prop_vacc}
