% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dhs_namibia_2013}
\alias{dhs_namibia_2013}
\title{2013 DHS vaccination data for Namibia}
\format{
Dataframe
}
\usage{
dhs_namibia_2013
}
\description{
Demographic and health data including region-level vaccination coverage from the 2013 Namibia Demographic and Health Survey.
}
\references{
\href{https://dhsprogram.com/pubs/pdf/FR298/FR298.pdf}{The  Nambia  Ministry  of  Health  and  Social  Services  (MoHSS)  and  ICF  International.  2014.  The  Namibia  Demographic and Health Survey 2013. Windhoek, Namibia, and Rockville, Maryland, USA: MoHSS and ICF International.}
}
\author{
John Giles
}
\keyword{datasets}
