% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_beta_params_age}
\alias{get_beta_params_age}
\title{Get Beta parameters for age groups}
\usage{
get_beta_params_age(age = NULL, vacc = NULL, breaks = NULL)
}
\arguments{
\item{age}{a vector giving the age at vaccination of each individual}

\item{vacc}{a binary vector indicating the vaccination status of each individual}

\item{breaks}{a scalar or vector of age group breaks (passed to \code{\link{.bincode}}). If NULL (default), calculates Beta parameters for all data together.}
}
\value{
A dataframe containing the sample size, mu, sigma, and Beta distribution paramters for each age group
}
\description{
This function finds the two shape parameters for the Beta distribution for aggregated age groups given individual-level
vaccination data. The individual-level data is comprised of the age of the individual (\code{age}) and a binary indicator of vaccination status
(\code{vacc}). The function first uses a Binomial GAM to estimate the mean (mu) and variance (sigma) of the proportion vaccinated for each
of the defined age groups and then finds the shape parameters of the Beta distribution analytically using \code{\link{get_beta_params}}.
}
\details{
Note that a Binomial GLM is used if the number of age groups is 2 or less.
}
\examples{
# dhs_namibia_2013 vaccination coverage file downloaded from the website
data("dhs_namibia_2013_namibia_2013")

# For all observations together (population mean)
tmp <- get_beta_params_age(age=dhs_namibia_2013$age.in.months,
                           vacc=dhs_namibia_2013$measles.y,
                           breaks=NULL)


# Two age groups < 12 months and > 12 months
tmp <- get_beta_params_age(age=dhs_namibia_2013$age.in.months,
                           vacc=dhs_namibia_2013$measles.y,
                           breaks=12)
par(mfrow=c(1,2))
for(i in 1:2) curve(dbeta(x, tmp$shape1[i], tmp$shape2[i]), 0, 1,
                    xlab='Proportion vaccinated', ylab='Density',
                    main=tmp$age[i], lwd=2)


# 6-month age groups
tmp <- get_beta_params_age(age=dhs_namibia_2013$age.in.months,
                           vacc=dhs_namibia_2013$measles.y,
                           breaks=seq(0, 60, 6))
par(mfrow=c(2,5))
for(i in 1:10) curve(dbeta(x, tmp$shape1[i], tmp$shape2[i]), 0, 1,
                     xlab='Proportion vaccinated', ylab='Density',
                     main=tmp$age[i], lwd=2)


# Each unique age in months
tmp <- get_beta_params_age(age=dhs_namibia_2013$age.in.months,
                           vacc=dhs_namibia_2013$measles.y,
                           breaks=seq(0, 60, 1))

par(mfrow=c(1,1))
plot(tmp$mu, type='l')


# 3-month age groups for each region
tmp <- foreach(i=unique(dhs_namibia_2013$region.residence), .combine='rbind') \%do\% {

  sel <- dhs_namibia_2013$region.residence == i

  cbind(
    data.frame(region=i),
    get_beta_params_age(age=dhs_namibia_2013$age.in.months[sel],
                        vacc=dhs_namibia_2013$measles.y[sel],
                        breaks=seq(0, 60, 3))
  )
}
}
\seealso{
Other beta_params: 
\code{\link{get_beta_params}()}
}
\author{
John Giles
}
\concept{beta_params}
